add test panic handler; for some reason isn't called in tests, main one is instead
